libreactor: Refactoring and optimization (#6219)
* libreactor: Remove the unused jansson dependency

* libreactor: move json() and plaintext() into helpers.c

* libreactor: move setup() to helpers.c and delete setup.c for the main (platform) implementation

* libreactor: replace setup.c with helpers.c for the libreactor-server implementation

* libreactor: run app as non-root user

* libreactor: use -march=native for app & libclo, and clean up other CFLAGS

libdynamic and libreactor already use -march=native by default

* libreactor: specify git commit hashes to ensure consistency

* libreactor: "optimize" server header

* libreactor: specify git commit hashes for libreactor-server as well

* libreactor: Unify both implementations under a single src folder

Share the same helpers rather than duplicating
