--- frameworks/C/libreactor/Makefile-server
@@ -1,11 +1,0 @@
-PROG    = libreactor
-OBJS    = src/setup.o src/main.o
-CFLAGS  = -std=gnu11 -Wall -Wextra -Wpedantic -O3 -g -flto -fcommon
-LDFLAGS = -pthread
-LDADD   = -lreactor -ldynamic -lclo
-
-$(PROG): $(OBJS)
-	$(CC) -o $@ $^ $(CFLAGS) $(LDFLAGS) $(LDADD)
-
-clean:
-	rm -f $(PROG) $(OBJS)
