+++ frameworks/C/libreactor/libreactor-server.dockerfile
@@ -7,42 +7,42 @@ WORKDIR /build
 
 ENV CC=gcc-10 AR=gcc-ar-10 NM=gcc-nm-10 RANLIB=gcc-ranlib-10
 
-RUN wget -q https://github.com/akheron/jansson/archive/v2.12.tar.gz -O jansson-2.12.tar.gz && \
-    tar xfz jansson-2.12.tar.gz && \
-    cd jansson-2.12 && \
-    autoreconf -fi && \
+RUN git clone https://github.com/fredrikwidlund/libdynamic && \
+    cd libdynamic && \
+    git checkout aee8f053c113 && \
+    ./autogen.sh && \
     ./configure && \
     make install
 
-RUN git clone https://github.com/fredrikwidlund/libdynamic && \
-    cd libdynamic && \
-    ./autogen.sh && \
-    ./configure --prefix=/usr CFLAGS="-Wall -Wextra -Wpedantic -O3 -g"&& \
-    make install
-
-# Using sed to remove the unused "#include <dynamic.h>" directive since it causes a build error: "unknown type name 'pthread_t'"
+# Remove the unused "#include <dynamic.h>" directive since it causes a build error: "unknown type name 'pthread_t'".
+# Specify configure prefix so libclo gets installed in /usr/lib like libdynamic and libreactor.
+# Add march=native. Both libdynamic and libreactor are already using it.
 RUN wget -q https://github.com/fredrikwidlund/libclo/releases/download/v1.0.0/libclo-1.0.0.tar.gz && \
     tar xfz libclo-1.0.0.tar.gz && \
     cd libclo-1.0.0 && \
     sed -i '/#include <dynamic.h>/d' ./src/clo.c && \
-    ./configure --prefix=/usr CFLAGS="-Wall -Wextra -Wpedantic -O3 -g" && \
+    ./configure --prefix=/usr CFLAGS="-march=native" && \
     make install
 
 RUN git clone https://github.com/fredrikwidlund/libreactor --single-branch --branch release-2.0 libreactor-2 && \
     cd libreactor-2 && \
+    git checkout 63fa717a8047 && \
     ./autogen.sh && \
-    ./configure --prefix=/usr CFLAGS="-Wall -Wextra -Wpedantic -O3 -g -fcommon" && \
+    ./configure && \
     make install
 
-COPY src-server/ /build/src/
-COPY Makefile-server /build/Makefile
+COPY src/ /build/src/
+COPY Makefile /build/Makefile
 
-RUN make
+RUN make libreactor-server
 
 
 FROM ubuntu:20.04
 
 WORKDIR /app
-COPY --from=builder /build/libreactor .
+COPY --from=builder /build/libreactor-server .
 
-CMD ["./libreactor"]
+RUN groupadd -r libreactor && useradd --no-log-init -r -g libreactor libreactor
+USER libreactor
+
+CMD ["./libreactor-server"]
