--- frameworks/C/libreactor/src-server/setup.h
@@ -1,6 +1,0 @@
-#ifndef SETUP_H_INCLUDED
-#define SETUP_H_INCLUDED
-
-void setup();
-
-#endif /* SETUP_H_INCLUDED */
