+++ frameworks/C/libreactor/src/libreactor.c
@@ -1,33 +1,35 @@
 #include <stdio.h>
-#include <stdlib.h>
 #include <stdint.h>
+#include <stdlib.h>
 #include <unistd.h>
-#include <string.h>
 #include <err.h>
 
 #include <dynamic.h>
 #include <reactor.h>
 #include <clo.h>
 
-#include "setup.h"
+#include "helpers.h"
+
 
 static core_status server_handler(core_event *event)
 {
-  static char json_msg[4096];
+  static char hello_string[] = "Hello, World!";
+  static char default_string[] = "Hello from libreactor!\n";
+  static clo_pair json_pair[] = {{ .string = "message", .value = { .type = CLO_STRING, .string = "Hello, World!" }}};
+  static clo json_object[] = {{ .type = CLO_OBJECT, .object = json_pair }};
 
   server *server = event->state;
   server_context *context = (server_context *) event->data;
 
   if (event->type == SERVER_REQUEST){
     if (segment_equal(context->request.target, segment_string("/json"))){
-      (void) clo_encode((clo[]) {clo_object({"message", clo_string("Hello, World!")})}, json_msg, sizeof(json_msg));
-      server_ok(context, segment_string("application/json"), segment_string(json_msg));
+      json(context, json_object);
     }
     else if (segment_equal(context->request.target, segment_string("/plaintext"))){
-      server_ok(context, segment_string("text/plain"), segment_string("Hello, World!"));
+      plaintext(context, hello_string);
     }
     else{
-      server_ok(context, segment_string("text/plain"), segment_string("Hello from libreactor!\n"));
+      plaintext(context, default_string);
     }
     return CORE_OK;
   }
